Executor: support container isolation for the persistent_worker
