+++ internal/executor/instance/persistentworker/persistentworker.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"github.com/cirruslabs/cirrus-ci-agent/api"
 	"github.com/cirruslabs/cirrus-cli/internal/executor/instance/abstract"
+	"github.com/cirruslabs/cirrus-cli/internal/executor/instance/persistentworker/isolation/container"
 	"github.com/cirruslabs/cirrus-cli/internal/executor/instance/persistentworker/isolation/none"
 	"github.com/cirruslabs/cirrus-cli/internal/executor/instance/persistentworker/isolation/parallels"
 	"github.com/cirruslabs/cirrus-cli/internal/logger"
@@ -28,6 +29,8 @@ func New(isolation *api.Isolation, logger logger.Lightweight) (abstract.Instance
 
 		return parallels.New(iso.Parallels.Image, iso.Parallels.User, iso.Parallels.Password,
 			strings.ToLower(iso.Parallels.Platform.String()), parallels.WithLogger(logger))
+	case *api.Isolation_Container_:
+		return container.New(iso.Container.Image, iso.Container.Cpu, iso.Container.Memory)
 	default:
 		return nil, fmt.Errorf("%w: unsupported isolation type %T", ErrInvalidIsolation, iso)
 	}
